Refactor for ai/setup.py client init
